+++ composer.json
@@ -16,19 +16,22 @@
         "phine/path": "~1.0",
         "nikic/php-parser": "~1.1",
         "monolog/monolog": "~1.13",
-        "react/react": "0.4.*",
         "phpdocumentor/reflection-docblock": "~2.0",
         "symfony/event-dispatcher": "^2.6",
         "doctrine/cache": "^1.4",
         "symfony/console": "^2.7",
         "php-di/php-di": "^5.1",
-        "php-di/symfony2-bridge": "^1.1"
+        "php-di/symfony2-bridge": "^1.1",
+        "amphp/file": "^0.1.2",
+        "amphp/socket": "^0.9.6"
     },
     "require-dev": {
         "behat/behat": "~3.0",
-        "peridot-php/peridot": "~1.15",
         "peridot-php/leo": "~1.4",
         "phpspec/prophecy": "^1.5",
-        "peridot-php/peridot-prophecy-plugin": "~1.0"
-    }
+        "peridot-php/peridot": "1.18",
+        "peridot-php/peridot-prophecy-plugin": "~1.0",
+        "phpspec/phpspec": "^2.5"
+    },
+    "minimum-stability": "beta"
 }
